@@ -1,5 +1,6 @@
 package net.b34tzepz.betterbuilding.block.custom;
 
+import com.mojang.serialization.MapCodec;
 import net.b34tzepz.betterbuilding.block.entity.ChairEntity;
 
 import net.minecraft.block.Block;
@@ -17,13 +18,13 @@ import net.minecraft.entity.decoration.ArmorStandEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemPlacementContext;
 import net.minecraft.item.ItemStack;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.state.StateManager;
-import net.minecraft.state.property.DirectionProperty;
+import net.minecraft.state.property.EnumProperty;
 import net.minecraft.state.property.Properties;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.BlockMirror;
 import net.minecraft.util.BlockRotation;
-import net.minecraft.util.Hand;
 import net.minecraft.util.function.BooleanBiFunction;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
@@ -32,19 +33,26 @@ import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
-import net.minecraft.world.WorldAccess;
+import net.minecraft.world.WorldView;
+import net.minecraft.world.tick.ScheduledTickView;
 import org.jetbrains.annotations.Nullable;
 import net.minecraft.util.math.random.Random;
 
 import java.util.stream.Stream;
 
 public class ChairBlock extends BlockWithEntity implements BlockEntityProvider {
-    public static final DirectionProperty FACING = Properties.HORIZONTAL_FACING;
+    public static final EnumProperty<Direction> FACING = Properties.HORIZONTAL_FACING;
+    public static final MapCodec<ChairBlock> CODEC = Block.createCodec(ChairBlock::new);
     public ArmorStandEntity stand = null;
 
     public ChairBlock(Settings settings) {
         super(settings);
         setDefaultState(this.stateManager.getDefaultState().with(Properties.HORIZONTAL_FACING, Direction.NORTH));
+    }
+
+    @Override
+    protected MapCodec<? extends BlockWithEntity> getCodec() {
+        return CODEC;
     }
 
     private static final VoxelShape NORTH_SHAPE =
@@ -114,17 +122,12 @@ public class ChairBlock extends BlockWithEntity implements BlockEntityProvider {
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
         Direction dir = state.get(FACING);
-        switch (dir) {
-            case SOUTH:
-                return SOUTH_SHAPE;
-
-            case WEST:
-                return WEST_SHAPE;
-
-            case EAST:
-                return EAST_SHAPE;
-        }
-        return NORTH_SHAPE;
+        return switch (dir) {
+            case SOUTH -> SOUTH_SHAPE;
+            case WEST -> WEST_SHAPE;
+            case EAST -> EAST_SHAPE;
+            default -> NORTH_SHAPE;
+        };
     }
 
     @Nullable
@@ -140,7 +143,7 @@ public class ChairBlock extends BlockWithEntity implements BlockEntityProvider {
 
 
     @Override
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+    protected ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
 
         if (player.getVehicle() == null) {
             stand = new ArmorStandEntity(EntityType.ARMOR_STAND, world);
@@ -166,16 +169,16 @@ public class ChairBlock extends BlockWithEntity implements BlockEntityProvider {
     }
 
     @Override
-    public void onStateReplaced(BlockState state, World world, BlockPos pos, BlockState newState, boolean moved) {
-        if (stand != null) stand.kill();
-        super.onStateReplaced(state, world, pos, newState, moved);
+    protected void onStateReplaced(BlockState state, ServerWorld world, BlockPos pos, boolean moved) {
+        if (stand != null) stand.kill(world);
+        super.onStateReplaced(state, world, pos, moved);
     }
 
     @Override
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         if (stand != null) {
             if (!stand.hasPlayerRider()) {
-                stand.kill();
+                stand.kill((ServerWorld) world);
                 stand = null;
             }
         }
@@ -183,13 +186,13 @@ public class ChairBlock extends BlockWithEntity implements BlockEntityProvider {
     }
 
     @Override
-    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction,
-                                                BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
+    protected BlockState getStateForNeighborUpdate(BlockState state, WorldView world, ScheduledTickView tickView, BlockPos pos, Direction direction,
+                                                   BlockPos neighborPos, BlockState neighborState, Random random) {
         BlockPos above = new BlockPos(pos.getX(), pos.getY() + 1, pos.getZ());
         if (!world.isAir(above)) {
             return Blocks.AIR.getDefaultState();
         }
-        return super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
+        return super.getStateForNeighborUpdate(state, world, tickView, pos, direction, neighborPos, neighborState, random);
     }
 
     @Override
