@@ -2,11 +2,12 @@ package net.b34tzepz.betterbuilding.block.custom;
 
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.Oxidizable;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
 
-public class OxidizablePillarBlock extends PillarBlock implements Oxidizable{
+public class OxidizablePillarBlock extends PillarBlock implements Oxidizable {
     private final net.minecraft.block.Oxidizable.OxidationLevel oxidationLevel;
 
     public OxidizablePillarBlock(net.minecraft.block.Oxidizable.OxidationLevel oxidationLevel, AbstractBlock.Settings settings) {
