@@ -10,7 +10,8 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.BlockView;
-import net.minecraft.world.WorldAccess;
+import net.minecraft.world.WorldView;
+import net.minecraft.world.tick.ScheduledTickView;
 import org.jetbrains.annotations.Nullable;
 
 public class PillarCoralBlock extends PillarBlock{
@@ -29,22 +30,21 @@ public class PillarCoralBlock extends PillarBlock{
     }
 
     @Override
-    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction, BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
+    protected BlockState getStateForNeighborUpdate(BlockState state, WorldView world, ScheduledTickView tickView, BlockPos pos, Direction direction, BlockPos neighborPos, BlockState neighborState, Random random) {
         if (!this.isInWater(world, pos)) {
-            world.scheduleBlockTick(pos, this, 60 + world.getRandom().nextInt(40));
+            tickView.scheduleBlockTick(pos, this, 60 + random.nextInt(40));
         }
-        return super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
+        return super.getStateForNeighborUpdate(state, world, tickView, pos, direction, neighborPos, neighborState, random);
     }
 
     protected boolean isInWater(BlockView world, BlockPos pos) {
-        if(world.getBlockState(pos).getBlock() instanceof PillarCoralBlock && world.getBlockState(pos).get(WATERLOGGED)){
-            return true;
-        }
         for (Direction direction : Direction.values()) {
             FluidState fluidState = world.getFluidState(pos.offset(direction));
-            if (!fluidState.isIn(FluidTags.WATER)) continue;
-            return true;
+            if (fluidState.isIn(FluidTags.WATER)) {
+                return true;
+            }
         }
+
         return false;
     }
 
