@@ -7,6 +7,7 @@ import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.registry.tag.EnchantmentTags;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
@@ -26,7 +27,7 @@ public class PillarIceBlock extends PillarBlock{
     @Override
     public void afterBreak(World world, PlayerEntity player, BlockPos pos, BlockState state, @Nullable BlockEntity blockEntity, ItemStack tool) {
         super.afterBreak(world, player, pos, state, blockEntity, tool);
-        if (EnchantmentHelper.getLevel(Enchantments.SILK_TOUCH, tool) == 0) {
+        if (!EnchantmentHelper.hasAnyEnchantmentsIn(tool, EnchantmentTags.PREVENTS_ICE_MELTING)) {
             if (world.getDimension().ultrawarm()) {
                 world.removeBlock(pos, false);
                 return;
@@ -41,7 +42,7 @@ public class PillarIceBlock extends PillarBlock{
 
     @Override
     public void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
-        if (world.getLightLevel(LightType.BLOCK, pos) > 11 - state.getOpacity(world, pos)) {
+        if (world.getLightLevel(LightType.BLOCK, pos) > 11 - state.getOpacity()) {
             this.melt(state, world, pos);
         }
     }
@@ -51,7 +52,7 @@ public class PillarIceBlock extends PillarBlock{
             world.removeBlock(pos, false);
         } else {
             world.setBlockState(pos, getMeltedState());
-            world.updateNeighbor(pos, getMeltedState().getBlock(), pos);
+            world.updateNeighbor(pos, getMeltedState().getBlock(), null);
         }
     }
 }
