@@ -16,6 +16,8 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldAccess;
+import net.minecraft.world.WorldView;
+import net.minecraft.world.tick.ScheduledTickView;
 
 public class PillarMagmaBlock extends PillarBlock{
     public PillarMagmaBlock(Settings settings) {
@@ -24,9 +26,10 @@ public class PillarMagmaBlock extends PillarBlock{
 
     @Override
     public void onSteppedOn(World world, BlockPos pos, BlockState state, Entity entity) {
-        if (!entity.bypassesSteppingEffects() && entity instanceof LivingEntity && !EnchantmentHelper.hasFrostWalker((LivingEntity)entity)) {
-            entity.damage(world.getDamageSources().hotFloor(), 1.0f);
+        if (!entity.bypassesSteppingEffects() && entity instanceof LivingEntity) {
+            entity.serverDamage(world.getDamageSources().hotFloor(), 1.0F);
         }
+
         super.onSteppedOn(world, pos, state, entity);
     }
 
@@ -36,11 +39,12 @@ public class PillarMagmaBlock extends PillarBlock{
     }
 
     @Override
-    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction, BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
+    protected BlockState getStateForNeighborUpdate(BlockState state, WorldView world, ScheduledTickView tickView, BlockPos pos, Direction direction, BlockPos neighborPos, BlockState neighborState, Random random) {
         if (direction == Direction.UP && neighborState.isOf(Blocks.WATER)) {
-            world.scheduleBlockTick(pos, this, 20);
+            tickView.scheduleBlockTick(pos, this, 20);
         }
-        return super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
+
+        return super.getStateForNeighborUpdate(state, world, tickView, pos, direction, neighborPos, neighborState, random);
     }
 
     @Override
