@@ -11,18 +11,22 @@ import net.minecraft.entity.projectile.ProjectileEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.stat.Stats;
 import net.minecraft.state.StateManager;
 import net.minecraft.state.property.BooleanProperty;
 import net.minecraft.state.property.Properties;
+import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
+import net.minecraft.world.GameRules;
 import net.minecraft.world.World;
+import net.minecraft.world.block.WireOrientation;
 import net.minecraft.world.event.GameEvent;
 import net.minecraft.world.explosion.Explosion;
 import org.jetbrains.annotations.Nullable;
@@ -47,7 +51,7 @@ public class PillarTntBlock extends PillarBlock{
     }
 
     @Override
-    public void neighborUpdate(BlockState state, World world, BlockPos pos, Block sourceBlock, BlockPos sourcePos, boolean notify) {
+    protected void neighborUpdate(BlockState state, World world, BlockPos pos, Block sourceBlock, @Nullable WireOrientation wireOrientation, boolean notify) {
         if (world.isReceivingRedstonePower(pos)) {
             primeTnt(world, pos);
             world.removeBlock(pos, false);
@@ -55,15 +59,16 @@ public class PillarTntBlock extends PillarBlock{
     }
 
     @Override
-    public void onBreak(World world, BlockPos pos, BlockState state, PlayerEntity player) {
-        if (!world.isClient() && !player.isCreative() && state.get(UNSTABLE)) {
+    public BlockState onBreak(World world, BlockPos pos, BlockState state, PlayerEntity player) {
+        if (!world.isClient() && !player.getAbilities().creativeMode && (Boolean)state.get(UNSTABLE)) {
             primeTnt(world, pos);
         }
-        super.onBreak(world, pos, state, player);
+
+        return super.onBreak(world, pos, state, player);
     }
 
     @Override
-    public void onDestroyedByExplosion(World world, BlockPos pos, Explosion explosion) {
+    public void onDestroyedByExplosion(ServerWorld world, BlockPos pos, Explosion explosion) {
         if (world.isClient) {
             return;
         }
@@ -73,47 +78,54 @@ public class PillarTntBlock extends PillarBlock{
         world.spawnEntity(tntEntity);
     }
 
-    public static void primeTnt(World world, BlockPos pos) {
-        primeTnt(world, pos, null);
+    public static boolean primeTnt(World world, BlockPos pos) {
+        return primeTnt(world, pos, null);
     }
 
-    private static void primeTnt(World world, BlockPos pos, @Nullable LivingEntity igniter) {
-        if (world.isClient) {
-            return;
+    private static boolean primeTnt(World world, BlockPos pos, @Nullable LivingEntity igniter) {
+        if (world instanceof ServerWorld serverWorld && serverWorld.getGameRules().getBoolean(GameRules.TNT_EXPLODES)) {
+            TntEntity tntEntity = new TntEntity(world, pos.getX() + 0.5, pos.getY(), pos.getZ() + 0.5, igniter);
+            world.spawnEntity(tntEntity);
+            world.playSound(null, tntEntity.getX(), tntEntity.getY(), tntEntity.getZ(), SoundEvents.ENTITY_TNT_PRIMED, SoundCategory.BLOCKS, 1.0F, 1.0F);
+            world.emitGameEvent(igniter, GameEvent.PRIME_FUSE, pos);
+            return true;
+        } else {
+            return false;
         }
-        TntEntity tntEntity = new TntEntity(world, (double)pos.getX() + 0.5, pos.getY(), (double)pos.getZ() + 0.5, igniter);
-        world.spawnEntity(tntEntity);
-        world.playSound(null, tntEntity.getX(), tntEntity.getY(), tntEntity.getZ(), SoundEvents.ENTITY_TNT_PRIMED, SoundCategory.BLOCKS, 1.0f, 1.0f);
-        world.emitGameEvent(igniter, GameEvent.PRIME_FUSE, pos);
     }
 
     @Override
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player2, Hand hand, BlockHitResult hit) {
-        ItemStack itemStack = player2.getStackInHand(hand);
-        if (itemStack.isOf(Items.FLINT_AND_STEEL) || itemStack.isOf(Items.FIRE_CHARGE)) {
-            primeTnt(world, pos, player2);
-            world.setBlockState(pos, Blocks.AIR.getDefaultState(), Block.NOTIFY_ALL | Block.REDRAW_ON_MAIN_THREAD);
-            Item item = itemStack.getItem();
-            if (!player2.isCreative()) {
-                if (itemStack.isOf(Items.FLINT_AND_STEEL)) {
-                    itemStack.damage(1, player2, player -> player.sendToolBreakStatus(hand));
+    protected ActionResult onUseWithItem(ItemStack stack, BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+        if (!stack.isOf(Items.FLINT_AND_STEEL) && !stack.isOf(Items.FIRE_CHARGE)) {
+            return super.onUseWithItem(stack, state, world, pos, player, hand, hit);
+        } else {
+            if (primeTnt(world, pos, player)) {
+                world.setBlockState(pos, Blocks.AIR.getDefaultState(), Block.NOTIFY_ALL_AND_REDRAW);
+                Item item = stack.getItem();
+                if (stack.isOf(Items.FLINT_AND_STEEL)) {
+                    stack.damage(1, player, LivingEntity.getSlotForHand(hand));
                 } else {
-                    itemStack.decrement(1);
+                    stack.decrementUnlessCreative(1, player);
                 }
+
+                player.incrementStat(Stats.USED.getOrCreateStat(item));
+            } else if (world instanceof ServerWorld serverWorld && !serverWorld.getGameRules().getBoolean(GameRules.TNT_EXPLODES)) {
+                player.sendMessage(Text.translatable("block.minecraft.tnt.disabled"), true);
+                return ActionResult.PASS;
             }
-            player2.incrementStat(Stats.USED.getOrCreateStat(item));
-            return ActionResult.success(world.isClient);
+
+            return ActionResult.SUCCESS;
         }
-        return super.onUse(state, world, pos, player2, hand, hit);
     }
 
     @Override
-    public void onProjectileHit(World world, BlockState state, BlockHitResult hit, ProjectileEntity projectile) {
-        if (!world.isClient) {
+    protected void onProjectileHit(World world, BlockState state, BlockHitResult hit, ProjectileEntity projectile) {
+        if (world instanceof ServerWorld serverWorld) {
             BlockPos blockPos = hit.getBlockPos();
             Entity entity = projectile.getOwner();
-            if (projectile.isOnFire() && projectile.canModifyAt(world, blockPos)) {
-                primeTnt(world, blockPos, entity instanceof LivingEntity ? (LivingEntity)entity : null);
+            if (projectile.isOnFire()
+                    && projectile.canModifyAt(serverWorld, blockPos)
+                    && primeTnt(world, blockPos, entity instanceof LivingEntity ? (LivingEntity)entity : null)) {
                 world.removeBlock(blockPos, false);
             }
         }
