@@ -1,7 +1,8 @@
 package net.b34tzepz.betterbuilding.block.entity;
 
 import net.b34tzepz.betterbuilding.item.inventory.ImplementedInventory;
-import net.b34tzepz.betterbuilding.screen.TeleporterBlockScreenHandler;
+import net.b34tzepz.betterbuilding.screen.custom.TeleporterBlockScreenHandler;
+import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerFactory;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
@@ -9,9 +10,10 @@ import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.Inventories;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.screen.ScreenHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.storage.ReadView;
+import net.minecraft.storage.WriteView;
 import net.minecraft.text.Text;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
@@ -20,9 +22,9 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 
-public class TeleporterBlockEntity extends BlockEntity implements NamedScreenHandlerFactory, ImplementedInventory {
+public class TeleporterBlockEntity extends BlockEntity implements ImplementedInventory, ExtendedScreenHandlerFactory<BlockPos> {
     private final DefaultedList<ItemStack> inventory= DefaultedList.ofSize(1, ItemStack.EMPTY);
-    public static ArrayList<BlockPos> teleporters = new ArrayList<BlockPos>();
+    public static ArrayList<BlockPos> teleporters = new ArrayList<>();
     public static int arraylength=0;
     public static int[] posX = new int[arraylength];
     public static int[] posY = new int[arraylength];
@@ -38,23 +40,11 @@ public class TeleporterBlockEntity extends BlockEntity implements NamedScreenHan
     }
 
     public Item getItem(){
-        return inventory.get(0).getItem();
+        return inventory.getFirst().getItem();
     }
 
     public static void tick(World world, BlockPos pos, BlockState state, TeleporterBlockEntity entity) {
 
-    }
-
-
-    @Override
-    public Text getDisplayName() {
-        return Text.of("Teleporter Block");
-    }
-
-    @Nullable
-    @Override
-    public ScreenHandler createMenu(int syncId, PlayerInventory inv, PlayerEntity player) {
-        return new TeleporterBlockScreenHandler(syncId, inv, this);
     }
 
     public void transferToArrays(){
@@ -73,31 +63,47 @@ public class TeleporterBlockEntity extends BlockEntity implements NamedScreenHan
     }
 
     public void transferFromArrays(){
-        teleporters = new ArrayList<BlockPos>();
+        teleporters = new ArrayList<>();
         for (int i = 0; i < arraylength; i++) {
             teleporters.add(new BlockPos(posX[i],posY[i],posZ[i]));
         }
     }
 
     @Override
-    protected void writeNbt(NbtCompound nbt) {
-        super.writeNbt(nbt);
-        Inventories.writeNbt(nbt, inventory);
-        nbt.putInt("arraylength",arraylength);
+    protected void writeData(WriteView view) {
+        super.writeData(view);
+        Inventories.writeData(view, inventory);
+        view.putInt("arraylength",arraylength);
         transferToArrays();
-        nbt.putIntArray("posX",posX);
-        nbt.putIntArray("posY",posY);
-        nbt.putIntArray("posZ",posZ);
+        view.putIntArray("posX",posX);
+        view.putIntArray("posY",posY);
+        view.putIntArray("posZ",posZ);
     }
 
     @Override
-    public void readNbt(NbtCompound nbt) {
-        Inventories.readNbt(nbt, inventory);
-        arraylength = nbt.getInt("arraylength");
-        posX=nbt.getIntArray("posX");
-        posY=nbt.getIntArray("posY");
-        posZ=nbt.getIntArray("posZ");
-        super.readNbt(nbt);
+    protected void readData(ReadView view) {
+        Inventories.readData(view, inventory);
+        arraylength = view.getInt("arraylength", arraylength);
+        posX=view.getOptionalIntArray("posX").orElse(new int[arraylength]);
+        posY=view.getOptionalIntArray("posY").orElse(new int[arraylength]);
+        posZ=view.getOptionalIntArray("posZ").orElse(new int[arraylength]);
+        super.readData(view);
         transferFromArrays();
+    }
+
+    @Override
+    public BlockPos getScreenOpeningData(ServerPlayerEntity player) {
+        return this.pos;
+    }
+
+    @Override
+    public Text getDisplayName() {
+        return Text.literal("Teleporter Block");
+    }
+
+    @Nullable
+    @Override
+    public ScreenHandler createMenu(int syncId, PlayerInventory inv, PlayerEntity player) {
+        return new TeleporterBlockScreenHandler(syncId, inv, this.pos);
     }
 }
