@@ -1,18 +1,18 @@
-package net.b34tzepz.betterbuilding.screen;
+package net.b34tzepz.betterbuilding.screen.custom;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.b34tzepz.betterbuilding.BetterBuilding;
+import net.minecraft.client.gl.RenderPipelines;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import net.minecraft.client.render.GameRenderer;
-import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
 public class TeleporterBlockScreen extends HandledScreen<TeleporterBlockScreenHandler> {
     private static final Identifier TEXTURE =
-            new Identifier(BetterBuilding.MOD_ID, "textures/gui/teleporter_block_gui.png");
+            Identifier.of(BetterBuilding.MOD_ID, "textures/gui/teleporter_block_gui.png");
 
     public TeleporterBlockScreen(TeleporterBlockScreenHandler handler, PlayerInventory inventory, Text title) {
         super(handler, inventory, title);
@@ -26,12 +26,9 @@ public class TeleporterBlockScreen extends HandledScreen<TeleporterBlockScreenHa
 
     @Override
     protected void drawBackground(DrawContext context, float delta, int mouseX, int mouseY) {
-        RenderSystem.setShader(GameRenderer::getPositionTexProgram);
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.setShaderTexture(0, TEXTURE);
         int x = (width - backgroundWidth) / 2;
         int y = (height - backgroundHeight) / 2;
-        context.drawTexture(TEXTURE, x, y, 0, 0, backgroundWidth, backgroundHeight);
+        context.drawTexture(RenderPipelines.GUI_TEXTURED, TEXTURE, x, y, 0, 0, backgroundWidth, backgroundHeight, 256, 256);
     }
 
     @Override
