@@ -1,6 +1,7 @@
-package net.b34tzepz.betterbuilding.screen;
+package net.b34tzepz.betterbuilding.screen.custom;
 
 import net.b34tzepz.betterbuilding.screen.slot.ModSymbolSlot;
+import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.Inventory;
@@ -8,21 +9,22 @@ import net.minecraft.inventory.SimpleInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.screen.ScreenHandler;
 import net.minecraft.screen.slot.Slot;
-import net.minecraft.world.World;
+import net.minecraft.util.math.BlockPos;
 
 public class TeleporterBlockScreenHandler extends ScreenHandler {
     private final Inventory inventory;
-    public TeleporterBlockScreenHandler(int syncId, PlayerInventory playerInventory) {
-        this(syncId, playerInventory, new SimpleInventory(1));
+
+    public TeleporterBlockScreenHandler(int syncId, PlayerInventory playerInventory, BlockPos pos) {
+        this(syncId, playerInventory, playerInventory.player.getWorld().getBlockEntity(pos));
     }
 
-    public TeleporterBlockScreenHandler(int syncId, PlayerInventory playerInventory, Inventory inventory) {
+    public TeleporterBlockScreenHandler(int syncId, PlayerInventory playerInventory, BlockEntity blockEntity) {
         super(ModScreenHandlers.TELEPORTER_BLOCK_SCREEN_HANDLER, syncId);
-        checkSize(inventory, 1);
-        this.inventory = inventory;
+        checkSize((Inventory)blockEntity, 1);
+        this.inventory = (Inventory)blockEntity;
         inventory.onOpen(playerInventory.player);
 
-        this.addSlot(new ModSymbolSlot(inventory, 0, 81, 36));
+        this.addSlot(new ModSymbolSlot(inventory, 0, 80, 35));
 
         addPlayerInventory(playerInventory);
         addPlayerHotbar(playerInventory);
@@ -61,14 +63,14 @@ public class TeleporterBlockScreenHandler extends ScreenHandler {
     private void addPlayerInventory(PlayerInventory playerInventory) {
         for (int i = 0; i < 3; ++i) {
             for (int l = 0; l < 9; ++l) {
-                this.addSlot(new Slot(playerInventory, l + i * 9 + 9, 8 + l * 18, 86 + i * 18));
+                this.addSlot(new Slot(playerInventory, l + i * 9 + 9, 8 + l * 18, 84 + i * 18));
             }
         }
     }
 
     private void addPlayerHotbar(PlayerInventory playerInventory) {
         for (int i = 0; i < 9; ++i) {
-            this.addSlot(new Slot(playerInventory, i, 8 + i * 18, 144));
+            this.addSlot(new Slot(playerInventory, i, 8 + i * 18, 142));
         }
     }
 }
